fix(openapi): error handling for incorrect project and version flags (#411)
* test: fix version creation flow in nock

It doesn't make sense that the API would return v1.0.1 if that's the version the user is trying to create! So this fixes that. The test still passes ¯\_(ツ)_/¯

* fix: properly bubble up errors finding versions

also add a couple tests

* refactor: get rid of unnecessary catch statement

This is being caught by the caller function, so this isn't necessary. Also the tests yield the same result.

* refactor: another place with unnecessary catch statements

Tests still pass, so I think we're good here.