docs(GHA): small updates (#468)
* chore: rearrange GHA examples

* docs: fix link, expand version usage

* docs: fix frontmatter examples to reflect the latest

* docs: language tweaks, link fixes, emojis

* chore: grammar fix

* revert: use `main` branch for docs links instead

After thinking about it a little bit more, I think it makes more sense for the user to view the docs on `main`, versus on the latest tag. Our docs sync with every push to `main` and we should probably version them so they're better tied to version tags (i.e. they're only published when we bump versions), but I think this is the more intuitive solution for now.