test(openapi/validate): cleanup 🧹  (#520)
* chore(deps): install ora

* chore: add ora to dependabot and reformat

* feat: add spinner for file validation

* feat: initial registry upload function

* refactor: move ora opts into shared function

* refactor: use shared ora opts in new function also

* feat: first pass at rewriting openapi command to support new flow

- also i realphabetized the imports

* chore: making the spinners spin

* test: refactor existing tests to work with new flow

* test: fix error code in test

I tried running recreating this and the API responds with a 4xx, not a 5xx in this case.

* test: adding coverage for PUT endpoint error handling

* chore: move two tests under describe block

it was bugging me how we had two tests that were uncategorized and clearly fit under the "upload" describe block (definitely my bad)

* test: add coverage for if registry endpoint fails

* chore: fix typo in test

* chore: remove unnecessary error handling

This error handling was unnecessary, as shown in the test I added. The validation section above it flags this, so there theoretically shouldn't be any errors bundling that wouldn't already be flagged in the validation section above.

* chore: test coverage in ora opts

* chore: oops

* chore: spinner language

* test: remove unused prompt handler

* chore: add test TODO

after noodling with this for a bit I couldn't figure it out :disappointed: so adding a TODO so someone can hopefully write a test for this.