feat(openapi): auto-detect API definitions (#567)
* refactor: add pathFilter option to readdirRecursive

This filter is going to be used when passing in the filter from the `ignore` package.

* chore(deps): install `ignore`

* test: refactor test return pattern

* test: add check for invalid JSON

this was a weird edge case with the oas-normalize function that we should make sure to account for

* feat: first pass at file detection

* test(validate): updates to reflect new file detection

* test(openapi): updates to reflect new file detection

* test: add CI-specific test

* refactor: swap out enquirer for prompts

After struggling for hours with `enquirer` and trying to get a single test to run properly, I caved and went with `prompts`. I'm completely blown away by how simple it was to write a test and I'm so mad about it lol :sob:

Tagging a few PRs below that I tried to reference when writing `enquirer` tests:
https://github.com/enquirer/enquirer/issues/234
https://github.com/enquirer/enquirer/issues/284

* test: add coverage for prompt

* test: address some branching coverage

* docs: update README.md with new guidance

* docs: some callout and grammatical tweaks

I think these docs should be in third person when describing what the tool does (e.g. "rdme does this" vs. "we do this"). I think using first-person grammar is fine when describing our recommendations.

Also using the GitHub callout syntax, see: https://github.com/orgs/community/discussions/16925

* refactor: move readdirRecursive to its own file

* refactor: slight readability improvement

* chore: slight copy change

* docs: more readable callout

* chore: remove swagger.json from .gitignore

We're no longer creating this file in tests, so no need to ignore it.

* chore: smol if-block refactors

Co-Authored-By: Jon Ursenbach <jon@ursenba.ch>

* refactor: better TS type + default

Co-Authored-By: Jon Ursenbach <jon@ursenba.ch>

* chore: lint

* feat: ignore git in subdirectories too

* docs: clarify guidance around gitignore

* Update src/lib/prepareOas.ts

Co-authored-by: Jon Ursenbach <erunion@users.noreply.github.com>

Co-authored-by: Jon Ursenbach <jon@ursenba.ch>
Co-authored-by: Jon Ursenbach <erunion@users.noreply.github.com>