feat: major version tags (#682)
* refactor: rename files

This consolidates getNodeVersion into another file and renames the set-version-output function to have a .js extension so we don't need to have weird `npm run lint` carveouts for it

* test: more test coverage

* refactor: use core function for output setting

This pattern for setting outputs is now deprecated

* fix: whoops

* refactor: use semver to parse out major version

we were previously using a zero-deps approach but now that we need to install deps for the set-version-output step, we might as well use a safer approach here.

* refactor: add ability to grab major version + test

this will set up some upcoming tagging changes

* feat: create file for tagging version

* docs: small update

* Revert "refactor: add ability to grab major version + test"

This reverts commit 0fc0fe2036e69bc87487c06d34dfc618a5493266.

* revert(kinda): use semver major instead

* refactor: use vanilla imports where possible

simple-git isn't really being super helpful here, so we might as well use a vanilla node function instead. i struggled to write tests so not gonna use our `getPkgVersion` function either since importing directly from package.json is just fine.

* chore: do a clean install + build before version bump

* docs: update github docs URLs

* docs: update MAINTAINERS guide

i feel like this flow is better, tagging a beta tripped me up the first time

* feat: surface major version tag in GHA onboarding

we should surface the v8 tag in the github actions that we create. this required a decent refactor to consolidate our major version retrieval logic. a bunch of snapshots had to be updated too.

* feat: surface major version tag in our docs

* chore: small JSDoc update

* fix: don't add major tag if prerelease

* fix: cmon dude

* refactor: use logger