fix: fallback to current version if user doesn't enter one
kind of weird, but our PUT /version endpoint requires a version as both the path param and as a body param, even if you aren't changing the name of the version. and passing in an empty string for the latter will result in the endpoint shitting the bed. that's API v1 for ya, folks!