fix(versions): general refactors to handle flag edge cases (#861)
* fix(versions): general refactors to handle flag edge cases

* refactor: rename function, add JSDoc

* fix: don't prompt for deprecated flag in create command

major bummer that prompts.inject doesn't play nicely with the `type` values in the prompt definitions. figured this out by just playing around locally 😞

* fix: actually screw it y'all get deprecated too

* fix: consolidate option definitions

* fix: fallback to current version if user doesn't enter one

kind of weird, but our PUT /version endpoint requires a version as both the path param and as a body param, even if you aren't changing the name of the version. and passing in an empty string for the latter will result in the endpoint shitting the bed. that's API v1 for ya, folks!

* chore: comment typo