refactor: mark key as required flag
so this is a little weird... but technically we should tell the user this is required! the cool thing about our prerun hook is that it prompts the user to login if they miss it. because of this, the key will always be populated by the time we get to the command so this is more type-safe as well!