test: mock out git (#972)
## 🧰 Changes

Speeds up a few tests by mocking out `git` commands.

I was poking around to investigate a bug
(https://linear.app/readme-io/issue/RM-8336/childdocs-are-createdsyncd-before-parentdocs-in-rdme-cli)
and noticed some test timeouts when running them locally. It looks like
for a number of tests, we were making actual requests like `git remote`.
I don't understand why those can be so slow, they're nearly instant when
running them in a shell?

**my machine before**

![image](https://github.com/readmeio/rdme/assets/451488/a8390e49-0125-4183-b69b-d36dc4081332)

**my machine after**

![image](https://github.com/readmeio/rdme/assets/451488/9db59ee3-46e9-48b9-8589-90a128df8211)

With all the parallelization, and the fact that the slowest part seems
to be the e2e tests, the actual speed up on CI is negligible. But for
running individual test files, they go from 5+ seconds to <1 second.

## 🧬 QA & Testing

Tests should pass???