fix: bring back #1117 without breaking everything (#1120)
## 🧰 Changes

we hit a bit of a wild edge case where the github action was only
breaking in production builds due to how oclif loads the `package.json`
and the only way we caught it was with [this
failure](https://github.com/readmeio/rdme/actions/runs/12267183540/job/34226921877).

this PR reverts https://github.com/readmeio/rdme/pull/1119 (which in
turn brings back https://github.com/readmeio/rdme/pull/1117) with a
slight tweak in
https://github.com/readmeio/rdme/pull/1120/commits/f9461b430e2318c00731c4f75d1506cd593d4970
to do the following:
- make our import paths friendlier to github actions
- manually copy over the `package.json` to our `dist/` directory
whenever we run `npm run build`. TS was automatically handling this when
we were using JSON imports before but now it's not able to pick up on
the import so we have to copy it over ourselves.

additionally in
https://github.com/readmeio/rdme/pull/1120/commits/669cb4f2854ff775023eeec3895b70998c70f920,
i added a little check so we can catch these sorts of things better
going forward.

## 🧬 QA & Testing

Provide as much information as you can on how to test what you've done.