fix(openapi): adding support for OpenAPI 3.1.1, 3.0.4 and basic auth fetches (#1153)
| 🚥 Resolves https://github.com/readmeio/rdme/issues/1152 |
| :------------------- |

## 🧰 Changes

This upgrades `oas-normalize` and `oas` in order to pull in updates made
to our OpenAPI parser to support OpenAPI 3.0.4 and 3.1.1 definitions. I
have also made a fix in `oas-normalize` to support basic auth
credentials in URLs, which will fix #1152.

Also upgraded `nock` to v14 while I'm here because that dropped today.