fix(openapi/upload): improved handling for YAML files (#1154)
| 🚥 Resolves RM-11893, https://github.com/readmeio/rdme/issues/1142 |
| :------------------- |

## 🧰 Changes

When uploading an API definition, `rdme` converts all files to JSON,
which means that YAML files are uploaded in JSON but with the file
extension of `.yaml` or `.yml`. This will convert files originally in
YAML back to YAML before uploading to ReadMe.

## 🧬 QA & Testing

Upload a spec in YAML and confirm in GitHug that it's been stored in
YAML. Also confirm that JSON specs are uploaded in JSON as well.

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>