feat(openapi/upload): add `--confirm-overwrite` flag (#1168)
| 🚥 Resolves https://github.com/readmeio/rdme/issues/1167 |
| :------------------- |

## 🧰 Changes

If an automated system runs `openapi upload` on a pre-existing file,
then an overwrite prompt is shown. There is no way to skip this prompt
which can lock up automated systems.

This PR adds a flag to skip the prompt.

## 🧬 QA & Testing

Provide as much information as you can on how to test what you've done.

**Case 1 - confirm existing behavior**
- Upload a new API spec to readme using `openapi upload {spec}
--key={key} --slug={slug}`
- Try uploading the API spec again to trigger the overwrite prompt

**Case 2 - test new behavior**
- Upload a new API spec to readme using `openapi upload {spec}
--key={key} --slug={slug}`
- Try uploading the API spec again using `openapi upload {spec}
--key={key} --slug={slug} --overwrite` to skip the overwrite prompt

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>