ci: update `docs upload` owlfooding (#1178)
| 🚥 Resolves ISSUE_ID |
| :------------------- |

Ok take 2 updating the docs syncing command! Turns out I had a really
out of date version locally, so it was easier just to make a new branch
instead of trying to squash a bunch of merge conflicts.

This updates the docs syncing to use the new version of rdme now that
our main docs project is on Refactored.

Provide as much information as you can on how to test what you've done.

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>