refactor: consolidate doc frontmatter validation logic (#1233)
## 🧰 Changes

i'm about to open up another PR for RM-12535 that adds a new
`-confirm-autofixes` flag, and as part of that i got fed up and totally
DRY'd up our duplicated validation logic so it's now basically the same
between `docs:upload` and `docs:migrate`.

i also added a tiny enhancement to our frontmatter autofixing logic that
allows the `hidden` attribute to accept flags like `'true'` and
`'false'`.

## 🧬 QA & Testing

`docs:migrate` is still in dire need of more test coverage so i've been
testing it locally and confirmed that it's still working as expected.
`docs:upload` has great test coverage so if those tests pass we should
be in good shape.