docs: improved flag sorting (#1239)
## 🧰 Changes

reworks how the help screens + docs sort our flags so we can define the
order. noticed this in https://github.com/readmeio/rdme/pull/1238 now
that we're using `--branch` instead of `--version`.

as a general nit, i prefer the `--key` flag is always first, and then
the `--branch` (f.k.a. `--version`) flag is always second.