fix: remove `/version` API deprecation warnings (#1238)
## 🧰 Changes

`rdme` users are now seeing these deprecation warnings to do some
underlying ReadMe APIv2 changes:

```
⚠️  ReadMe API Warning: Our `/versions` endpoint URI family has been renamed to `/branches` and will be removed in the future.
```

with the changes in this PR, the documented flag to use for the `openapi
upload` and `docs upload` command is `--branch`. the `--version` flag
will continue to work, but users will now see this warning[^1]:

```
›   Warning: The "--version" flag has been deprecated. Use "--branch" instead.
```

## 🧬 QA & Testing

added a bunch of test coverage to account for this and also updated our
docs throughout. please flag any discrepancies you see!

[^1]: i know, it's on my list to consolidate our warning formatting 🫠