fix(openapi/upload): problem with `FormData` compositions on Node 24 (#1241)
| 🚥 Resolves CX-1950 |
| :------------------- |

## 🧰 Changes

Undici changed something with how streams are processed in `FormData`
payloads in the version that ships with Node 24 that has broken OpenAPI
definition uploads within rdme there. Because we don't _really_ need to
send these API definitions as a stream in our `multipart/form-data`
payload to our public API I've changed them to just send it as a regular
`File` object.

## 🧬 QA & Testing

Tested this locally on Node 22 and 24 and everything is green.