test: consolidate test suite for all markdown syncing commands (#1262)
## 🧰 Changes

this piggybacks off of @emilyskuo's work in
https://github.com/readmeio/rdme/pull/1261 and implements a suggestion i
had in [this
comment](https://github.com/readmeio/rdme/pull/1261/files#r2151162521)
(i.e., consolidating the tests for guides and reference uploads into a
single test suite wrapped in a `describe.each`). for now it consolidates
the tests for `rdme docs upload` and `rdme reference upload`, but i
figured it'll alleviate some of the burden of shipping support for
custom pages, etc. going forward.

as part of this, i did the following:

- [x] wrapped everything in a `describe.each`
- [x] added a bunch of template strings for the routes
- [x] duplicated a handful of fixtures
- [x] incorporated this feedback:
https://github.com/readmeio/rdme/pull/1261/files#r2151151087

## 🧬 QA & Testing

i went through all of the snapshots to make sure it all checks out, but
please feel free to look through everything yourself!