feat: new `reference upload` command (#1261)
| 🚥 Resolves RM-12965 |
| :------------------- |

## 🧰 Changes

Creating a new `reference upload` command for users to upload MD pages
to the reference section. This is entirely based on the `docs upload`
command, with most of it being exactly the same, minus the route and
some typing.

I had looked into making this an alias of `docs upload`, but oclif
doesn't tell you how the command was invoked, and we'd have to parse
`argv`s to figure out if docs or reference, which could potentially work
but would also be pretty brittle. Otherwise, we could create a general
upload command for markdown pages and add a type arg, which would be
changing up behavior even more.

## 🧬 QA & Testing

@kanadgupta consolidated our markdown upload command tests in
https://github.com/readmeio/rdme/pull/1262 🎉

---------

Co-authored-by: Kanad Gupta <git@kanad.dev>