fix(pages/upload): decouple autofixing and uploads (#1282)
| 🚥 Resolves RM-12355 |
| :------------------- |

## 🧰 Changes

with the way our `docs upload` / `reference upload` / etc. commands are
set up, this is how a command workflow might go with what’s in `next`:

1. user runs command
2. `rdme` detects autofixable frontmatter, prompts user to fix it
3. user says yes
4. frontmatter is autofixed
5. pages are then uploaded to ReadMe

based on internal feedback, that flow contains too _**little**_ friction
(e.g., the idea of doing steps 4 and 5 above in one fell swoop is a
little heavy-handed and is throwing folks off).

this PR tweaks this a bit so this is the new flow:

1. user runs command
2. `rdme` detects autofixable frontmatter, prompts user to fix it
3. user says yes
4. frontmatter is autofixed
5. `rdme` displays frontmatter autofixing results, prompts user to run
the command again in order to upload files, and then exits.
6. user runs `rdme` again, per instructions
7. pages are then uploaded to ReadMe

## 🧬 QA & Testing

i still need to QA this against a local project (will mark this as
officially ready for review when that's done), but the tests around this
are pretty airtight.

**edit**: QA is done! here's what the new flow looks like:

![CleanShot 2025-07-02 at 14 08
37@2x](https://github.com/user-attachments/assets/d71e0f20-d8be-464f-825e-88358b76b480)

---------

Co-authored-by: Jon Ursenbach <erunion@users.noreply.github.com>
Co-authored-by: Emily Kuo <58803587+emilyskuo@users.noreply.github.com>