refactor(openapi): cleaner approach to `prepareOas` helper (#1312)
## 🧰 Changes

as part of the work in https://github.com/readmeio/rdme/pull/1313, this
PR refactors our `prepareOas` helper (and by extension, the entire
`openapi` family of commands) to be a little bit more conformant to the
oclif command class pattern. this will clean up our debug logs a bit and
should make it a bit easier to add new `openapi` commands going forward.

## 🧬 QA & Testing

no end user changes — if tests continue to pass we should be in good
shape!