fix(openapi/upload): custom slugs for URL uploads (#1313)
| 🚥 Resolves CX-1881 |
| :------------------- |

## 🧰 Changes

users that passed in a URL into `openapi upload` with a custom slug
(e.g., `rdme openapi upload https://example.com/petstore.json --slug
custom-slug`) were seeing confusing behavior because the custom slug
isn't reflected in readme. this is a limitation with our API, where the
slug is inferred for file uploads but there's no parameter for custom
slugs.

we'll keep an eye on customer interest for the APIv2 support for such a
parameter (my gut says it's not worth it just b/c it's additional
complexity in our API and the workaround is trivial). in the interim,
this PR tweaks the logic so we send a fetched/bundled version of the
spec as a file payload to the API, as opposed to sending the URL itself.

## 🧬 QA & Testing

do tests pass?