feat(plugins): add `runRdmeCommand` helper (#1292)
## 🧰 Changes

as part of the experimental plugin work we did a few months ago, i wrote
up this little helper method to have a type-safe way to invoke `rdme`
commands. this PR adds that method in. might be good to have for future
plugin-type work!

## 🧬 QA & Testing

i confirmed that this method works as expected locally! also added a
unit test for this.