fix(openapi/upload): better handling for mismatching files/slugs (#1345)
| 🚥 Resolves CX-2293 |
| :------------------- |

## 🧰 Changes

we previously had a restriction in place where you could not upload a
JSON file (e.g., `petstore.json`) to a slug with a YAML file extension
(e.g., `--slug file.yml`). the reasoning was that we didn't want to
create any unexpected client-side behaviors and we wanted to prioritize
explicitness.

however, we've been getting feedback that it's misleading how `rdme`
directs customers towards certain slugs, only to see an error that they
can't use that slug due to this file extension restriction.

this PR lifts that restriction and makes it so `rdme` will convert a
JSON file to YAML prior to uploading if the slug has a YAML file
extension (or vice versa). it will also emit a warning to the user when
this happens!

## 🧬 QA & Testing

added a bunch of tests for this behavior!

---------

Co-authored-by: Emily Kuo <58803587+emilyskuo@users.noreply.github.com>